Got rid of old classes (now alchemy only) and added workaround for sqlite_master table.
