+++ tests/testMovieDbClasses.py
@@ -21,6 +21,7 @@ from movieDbClasses import Titles
 from movieDbClasses import TitlesInLocations
 from movieDbClasses import Shows
 from movieDbClasses import Translations
+from movieDbClasses import SQLite_Master
 
 class testMovieDbClasses(unittest.TestCase):
     """
@@ -42,7 +43,8 @@ class testMovieDbClasses(unittest.TestCase):
             Titles,
             TitlesInLocations,
             Shows,
-            Translations
+            Translations,
+            SQLite_Master
         ]
 
     _expectedColumns = {
@@ -79,6 +81,13 @@ class testMovieDbClasses(unittest.TestCase):
                 "lang_to",
                 "title_from_ref",
                 "title_to_ref",
+                ],
+            "SQLite_Master": [
+                "type",
+                "name",
+                "table_name",
+                "root_page",
+                "sql",
                 ]
         }
 
@@ -111,9 +120,7 @@ class testMovieDbClasses(unittest.TestCase):
         unittest.TestCase.tearDown(self)
 
     def testDbClasses(self):
-        """
-        Verifies that the package defines the expected classes (one per table).
-        """
+        """Verifies that the package defines the expected classes (one per table)."""
         missingClasses = []
 
         self.logger.info("*** Checking the classes in the package...")
@@ -138,9 +145,7 @@ class testMovieDbClasses(unittest.TestCase):
         self.logger.info("Package check passed.")
 
     def testDbClassesColumns(self):
-        """
-        Verifies that the definitions of the db classes conform to the expectations.
-        """
+        """Verifies that the definitions of the db classes conform to the expectations."""
         self.logger.info("*** Verifying db classes definitions...")
 
         missingColumns = {}
@@ -149,7 +154,7 @@ class testMovieDbClasses(unittest.TestCase):
         # Check all the classes in the module
         allClasses = inspect.getmembers(movieDbClasses, inspect.isclass)
 
-        for clazz in [x[0] for x in allClasses if repr(x[1])[8:].startswith("movieDbClasses") and not repr(x[1])[8:].endswith("movieDbBaseClass'>")]:
+        for clazz in [x[0] for x in allClasses if repr(x[1]).find("SQLite_Master") == -1 and repr(x[1])[8:].startswith("movieDbClasses") and not repr(x[1])[8:].endswith("movieDbBaseClass'>")]:
             self.logger.info("Testing class %s", clazz)
 
             # 1. create one instance of that class
@@ -190,9 +195,7 @@ class testMovieDbClasses(unittest.TestCase):
             self.logger.info("Columns check passed.")
 
     def testInsertSites(self):
-        """
-        Verifies the 1-to-many relationship between Locations and sites works.
-        """
+        """Verifies the 1-to-many relationship between Locations and sites works."""
         self.logger.info("Testing creation of Locations and Sites")
 
         self.logger.debug("\tCreating locations.")
@@ -227,9 +230,7 @@ class testMovieDbClasses(unittest.TestCase):
         self.assertTrue(aSite2.id != None, "aSite2 should have been committed correctly and have a valid id.")
 
     def testInsertTitlesInLocations(self):
-        """
-        Verifies the 1-to-many relationship between Locations and sites works.
-        """
+        """Verifies the 1-to-many relationship between Locations and sites works."""
         self.logger.info("Testing creation of Locations and Titles(InLocation)")
 
         self.logger.debug("\tCreating locations.")
